requirements.txt
streamlit
Pillow
requests
google-generativeai
openai
streamlit-component-lib
anthropic
openrouter